Move pagination to its own file
