+++ publicapi/publicapi.go
@@ -2,12 +2,9 @@ package publicapi
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"github.com/gin-gonic/gin"
-	"time"
-
 	db "github.com/mikeydub/go-gallery/db/gen/coredb"
 	"github.com/mikeydub/go-gallery/event"
 
@@ -171,60 +168,4 @@ func pushFeedEvent(ctx context.Context, evt db.Event) {
 		logger.For(ctx).Error(err)
 		sentryutil.ReportError(ctx, err)
 	}
-}
-
-func encodeTimeIDCursor(t time.Time, id persist.DBID) (string, error) {
-	timeBytes, err := t.MarshalBinary()
-	if err != nil {
-		return "", err
-	}
-
-	idBytes := []byte(id)
-
-	bytes := make([]byte, 0, 1+len(timeBytes)+len(idBytes))
-
-	// The first byte declares how many bytes the time.Time is
-	bytes = append(bytes, byte(len(timeBytes)))
-
-	bytes = append(bytes, timeBytes...)
-	bytes = append(bytes, idBytes...)
-
-	return base64.StdEncoding.EncodeToString(bytes), nil
-}
-
-func decodeTimeIDCursor(cursor string) (time.Time, persist.DBID, error) {
-	bytes, err := base64.StdEncoding.DecodeString(cursor)
-	if err != nil || len(bytes) == 0 {
-		return time.Time{}, "", errBadCursorFormat
-	}
-
-	// The first byte declares how many bytes the time.Time is
-	timeLen := int(bytes[0])
-	if timeLen > len(bytes)-1 {
-		return time.Time{}, "", errBadCursorFormat
-	}
-
-	t := time.Time{}
-	err = t.UnmarshalBinary(bytes[1 : timeLen+1])
-	if err != nil {
-		return time.Time{}, "", err
-	}
-
-	id := persist.DBID("")
-
-	// It's okay for the ID to be empty
-	if len(bytes) > (timeLen + 1) {
-		id = persist.DBID(bytes[timeLen+1:])
-	}
-
-	return t, id, nil
-}
-
-type PageInfo struct {
-	Total           *int
-	Size            int
-	HasPreviousPage bool
-	HasNextPage     bool
-	StartCursor     string
-	EndCursor       string
 }
